EDSC-4543: Renaming files
