+++ static/src/js/components/Map/LayerPicker/LayerPicker.scss
@@ -1,4 +1,4 @@
-@use '../../../css/utils';
+@use '../../../../css/utils';
 
 .layer-picker {
   &__layers {
