+++ static/src/js/components/Map/Legend/Legend.scss
@@ -1,4 +1,4 @@
-@use '../../../css/utils';
+@use '../../../../css/utils';
 
 .legend {
   position: relative;
