+++ static/src/js/components/Map/LegendControl/LegendControl.scss
@@ -1,4 +1,4 @@
-@use '../../../css/utils';
+@use '../../../../css/utils';
 
 .legend-control {
   position: absolute;
